# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API",
        "description": "API for school management system including user authentication, and resource management",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/tracks": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Get all tracks",
                "operationId": "ad1baacaa4dfb415363918e8e58e0833",
                "responses": {
                    "200": {
                        "description": "List of all tracks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tracks": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "track_name": {
                                                        "type": "string",
                                                        "example": "Academic"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tracks retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Create a new track",
                "operationId": "844a03e0598aeb45b5d45dd34dcf92d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "track_name"
                                ],
                                "properties": {
                                    "track_name": {
                                        "type": "string",
                                        "example": "Academic"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Track created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "track": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "track_name": {
                                                    "type": "string",
                                                    "example": "Academic"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Track created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tracks/{id}": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Get a specific track",
                "operationId": "8fc9f230b412ad82102ca0395758fc19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Track ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Track details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "track": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "track_name": {
                                                    "type": "string",
                                                    "example": "Academic"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Track retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Track not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}